+++ src/Components/HelpInfo.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+import { Help } from "../Styles";
+import Tooltip from "./Tooltip";
+
+function HelpInfo({ children }) {
+  return (
+    <Tooltip content={children}>
+      <Help>i</Help>
+    </Tooltip>
+  );
+}
+
+export default React.memo(HelpInfo);